added pulses
